Refactor: DB 디렉토리 추가 및 파일 위치변경
